backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]
httpx==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis and caching
redis==5.0.1
celery==5.3.4

# Data processing and ML
pandas==2.1.3
numpy==1.25.2
jupyter-client==9.0.1
ipykernel==6.26.0

# AI and LLM Providers
openai==1.3.7
google-generativeai==0.3.2
ollama==0.1.7
langchain==0.0.350
langchain-openai==0.0.2
langchain-google-genai==0.0.6
langchain-community==0.0.10

# File handling and storage
boto3==1.34.0
minio==7.2.0
python-magic==0.4.27

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
websockets==12.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Monitoring and logging
structlog==23.2.0
prometheus-client==0.19.0 